+++ Bit_Team_Project/effect.py
@@ -1,3 +1,7 @@
 # 스킬 이펙트 모아서 여기 저장
 # 예시
 # 스킬 이펙트 이름 = ["스킬 이미지 주소 1", "스킬 이미지 주소 2" ...]
+test_skill_effect = ["Skill/Bomb/bomb_effect/1.png", "Skill/Bomb/bomb_effect/2.png", "Skill/Bomb/bomb_effect/3.png", "Skill/Bomb/bomb_effect/4.png", "Skill/Bomb/bomb_effect/5.png", 
+"Skill/Bomb/bomb_effect/6.png", "Skill/Bomb/bomb_effect/7.png", "Skill/Bomb/bomb_effect/8.png"]
+test_skill_icon = "Sprite/enemy.png"
+skill_throwing_test = "Skill/Bomb/bomb.png"
